vsr mv basis field render

+++ vsr/vsr_basis.h
@@ -10,7 +10,7 @@ namespace vsr{
 using namespace std;
 	
 
-typedef double VT; 
+typedef float VT; 
 typedef short TT;
 
 template<int N>
